Score

+++ 0x0C-MySQL_Advanced/6-bonus.sql
@@ -1,7 +1,7 @@
 -- Write a SQL script that creates a stored procedure AddBonus that adds a new correction for a student.
 DELIMITER //
 
-CREATE PROCEDURE AddBonus(user_id INT, project_name varchar(255), score FLOAT)
+CREATE PROCEDURE AddBonus(user_id INT, project_name varchar(255), score INT)
 BEGIN
 IF EXISTS(SELECT * FROM projects WHERE name = project_name LIMIT 1)
 THEN SET @projectid = (SELECT id FROM projects WHERE name = project_name LIMIT 1);
